Add Chef class to orchestrate complex actions
Some of the code is currently tied to the Storage driver, whereas it interacts
with many other components. This should be moved out.
That patch introduces the Chef, a class responsible for doing complex actions
that require interactions and orchestrations with different actors.
