+++ gnocchi/tests/test_chef.py
@@ -1,0 +1,57 @@
+# -*- encoding: utf-8 -*-
+#
+# Copyright © 2018 Red Hat
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+import datetime
+
+import numpy
+
+from gnocchi import chef
+from gnocchi import incoming
+from gnocchi import indexer
+from gnocchi.tests import base
+
+
+def datetime64(*args):
+    return numpy.datetime64(datetime.datetime(*args))
+
+
+class TestChef(base.TestCase):
+    def setUp(self):
+        super(TestChef, self).setUp()
+        self.metric, __ = self._create_metric()
+        self.chef = chef.Chef(self.coord, self.incoming,
+                              self.index, self.storage)
+
+    def test_delete_nonempty_metric_unprocessed(self):
+        self.incoming.add_measures(self.metric.id, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+        ])
+        self.index.delete_metric(self.metric.id)
+        self.trigger_processing()
+        __, __, details = self.incoming._build_report(True)
+        self.assertIn(str(self.metric.id), details)
+        self.chef.expunge_metrics(sync=True)
+        __, __, details = self.incoming._build_report(True)
+        self.assertNotIn(str(self.metric.id), details)
+
+    def test_delete_expunge_metric(self):
+        self.incoming.add_measures(self.metric.id, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+        ])
+        self.trigger_processing()
+        self.index.delete_metric(self.metric.id)
+        self.chef.expunge_metrics(sync=True)
+        self.assertRaises(indexer.NoSuchMetric, self.index.delete_metric,
+                          self.metric.id)
